move language detection into modelds utils and the tests
