+++ doctr/models/predictor/tensorflow.py
@@ -9,11 +9,10 @@ import numpy as np
 import tensorflow as tf
 
 from doctr.io.elements import Document
-from doctr.models._utils import estimate_orientation
+from doctr.models._utils import estimate_orientation, get_language
 from doctr.models.detection.predictor import DetectionPredictor
 from doctr.models.recognition.predictor import RecognitionPredictor
 from doctr.utils.geometry import rotate_boxes, rotate_image
-from doctr.utils.lang_detect import get_language
 from doctr.utils.repr import NestedObject
 
 from .base import _OCRPredictor
